Support check_mode in katello_manifest

+++ modules/katello_manifest.py
@@ -115,8 +115,9 @@ def manifest(module, organization, state, manifest_path=None, redhat_repository_
     manifest_present = current_manifest is not None
 
     if organization.redhat_repository_url != redhat_repository_url:
-        organization.redhat_repository_url = redhat_repository_url
-        organization.update({'redhat_repository_url'})
+        if not module.check_mode():
+            organization.redhat_repository_url = redhat_repository_url
+            organization.update({'redhat_repository_url'})
         changed = True
 
     if state == 'present':
@@ -125,7 +126,8 @@ def manifest(module, organization, state, manifest_path=None, redhat_repository_
                 files = {'content': (manifest_path, manifest_file, 'application/zip')}
                 data = {'organization_id': organization.id, 'repository_url': redhat_repository_url}
                 headers = {'content_type': 'multipart/form-data', 'multipart': 'true'}
-                Subscription().upload(data=data, files=files, headers=headers)
+                if not module.check_mode():
+                    Subscription().upload(data=data, files=files, headers=headers)
                 changed = True
         except IOError as e:
             module.fail_json(msg="Unable to open the manifest file: %s" % e)
@@ -137,13 +139,15 @@ def manifest(module, organization, state, manifest_path=None, redhat_repository_
             else:
                 module.fail_json(msg="Upload of the mainfest failed: %s" % e)
     elif state == 'absent' and manifest_present:
-        Subscription().delete_manifest(data={'organization_id': organization.id})
+        if not module.check_mode():
+            Subscription().delete_manifest(data={'organization_id': organization.id})
         changed = True
     elif state == 'refreshed':
         if not manifest_present:
             module.fail_json(msg="No manifest found to refresh.")
         else:
-            Subscription().refresh_manifest(data={'organization_id': organization.id})
+            if not module.check_mode():
+                Subscription().refresh_manifest(data={'organization_id': organization.id})
             changed = True
     return changed
 
@@ -162,7 +166,8 @@ def main():
         ),
         required_if=[
             ['state', 'present', ['manifest_path']],
-        ]
+        ],
+        supports_check_mode=True,
     )
 
     if not HAS_NAILGUN_PACKAGE:
